+++ src/src/commands/subscribeCommand.ts
@@ -1,7 +1,7 @@
 import {SlashCommandBuilder, SlashCommandSubcommandBuilder} from '@discordjs/builders';
 import {CommandInteraction} from 'discord.js';
 import {AbstractCommand} from './abstractCommand';
-import {SubscriptionType, ZKillSubscriber} from '../zKillSubscriber';
+import {LimitType, SubscriptionType, ZKillSubscriber} from '../zKillSubscriber';
 
 export class SubscribeCommand extends AbstractCommand {
     protected name = 'zkill-subscribe';
@@ -15,16 +15,43 @@ export class SubscribeCommand extends AbstractCommand {
         const subCommand = interaction.options.getSubcommand(true) as SubscriptionType;
         const id = interaction.options.getNumber('id', false);
         const minValue = interaction.options.getNumber('min-value');
-        sub.subscribe(subCommand, interaction.guildId, interaction.channelId, id ? id : undefined, minValue ? minValue : 0);
+        const limitRegion = interaction.options.getString('limit-region-ids');
+        const limitConstellation = interaction.options.getString('limit-constellation-ids');
+        const limitSystem = interaction.options.getString('limit-system-ids');
 
         let reply = 'We subscribed to zkillboard channel: ' + interaction.options.getSubcommand();
+        let limitType: LimitType = LimitType.NONE, limitIds;
+        if(limitConstellation || limitRegion || limitSystem) {
+            if(limitConstellation && limitRegion || limitConstellation && limitSystem || limitRegion && limitSystem) {
+                interaction.reply({content: 'Only one type of limit is allowed!', ephemeral: true});
+                return;
+            }
+            if(limitRegion) {
+                limitType = LimitType.REGION;
+                limitIds = limitRegion;
+                reply = 'Region filter: + ' + limitRegion;
+            }
+            if(limitConstellation) {
+                limitType = LimitType.CONSTELLATION;
+                limitIds = limitConstellation;
+                reply = 'Constellation filter: + ' + limitRegion;
+            }
+            if(limitSystem) {
+                limitType = LimitType.SYSTEM;
+                limitIds = limitSystem;
+                reply = 'System filter: + ' + limitRegion;
+
+            }
+        }
+        sub.subscribe(subCommand, interaction.guildId, interaction.channelId, id ? id : undefined, minValue ? minValue : 0, limitType, limitIds);
+
         if(id) {
             reply += ' ID: ' + id;
         }
         if(minValue) {
             reply += ' Min Value: ' + minValue.toLocaleString('en');
         }
-        interaction.reply(reply);
+        interaction.reply({content: reply, ephemeral: true});
     }
 
     getCommand(): SlashCommandBuilder {
@@ -39,6 +66,21 @@ export class SubscribeCommand extends AbstractCommand {
                     .setDescription('ID for the corporation')
                     .setRequired(true)
             )
+            .addStringOption(option =>
+                option.setName('limit-region-ids')
+                    .setDescription('Limit to region id, comma seperated ids')
+                    .setRequired(false)
+            )
+            .addStringOption(option =>
+                option.setName('limit-constellation-ids')
+                    .setDescription('Limit to constellation id, comma seperated ids')
+                    .setRequired(false)
+            )
+            .addStringOption(option =>
+                option.setName('limit-system-ids')
+                    .setDescription('Limit to system id, comma seperated ids')
+                    .setRequired(false)
+            )
             .addNumberOption(option =>
                 option.setName('min-value')
                     .setDescription('Minimum isk to show the entry')
@@ -51,6 +93,21 @@ export class SubscribeCommand extends AbstractCommand {
                 option.setName('id')
                     .setDescription('ID for the alliance')
                     .setRequired(true)
+            )
+            .addStringOption(option =>
+                option.setName('limit-region-ids')
+                    .setDescription('Limit to region id, comma seperated ids')
+                    .setRequired(false)
+            )
+            .addStringOption(option =>
+                option.setName('limit-constellation-ids')
+                    .setDescription('Limit to constellation id, comma seperated ids')
+                    .setRequired(false)
+            )
+            .addStringOption(option =>
+                option.setName('limit-system-ids')
+                    .setDescription('Limit to system id, comma seperated ids')
+                    .setRequired(false)
             )
             .addNumberOption(option =>
                 option.setName('min-value')
@@ -79,6 +136,21 @@ export class SubscribeCommand extends AbstractCommand {
                     .setDescription('ID for the region')
                     .setRequired(true)
 
+            )
+            .addStringOption(option =>
+                option.setName('limit-region-ids')
+                    .setDescription('Limit to region id, comma seperated ids')
+                    .setRequired(false)
+            )
+            .addStringOption(option =>
+                option.setName('limit-constellation-ids')
+                    .setDescription('Limit to constellation id, comma seperated ids')
+                    .setRequired(false)
+            )
+            .addStringOption(option =>
+                option.setName('limit-system-ids')
+                    .setDescription('Limit to system id, comma seperated ids')
+                    .setRequired(false)
             )
             .addNumberOption(option =>
                 option.setName('min-value')