+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+2022-06-30
+---
+
+- make alliance/corp/char killmails filterable by region/constellation/system (#5)
+- improved killmail embedings by fetching the open graph data from zkillboard url (#2)
+- colors for kills (green) / losses (red) on the embeds (#2)
+- __attention: new commands will roll out within 24 hours__
+
 2022-06-23
 ---